+++ src/extension.ts
@@ -4,6 +4,7 @@ import { headerComment } from './commands/headerComment';
 import { createProblem } from './commands/createProblem';
 import { pushToGithub } from './commands/pushToGithub';
 import { makeWorkflow } from './libs/makeWorkflow';
+import { showManual } from './commands/showManual';
 
 export function activate(context: vscode.ExtensionContext)
 {
@@ -13,8 +14,9 @@ export function activate(context: vscode.ExtensionContext)
 	const author = config.get<string>('author', '');
 	if (extension === undefined || author === undefined || extension === '' || author === '')
 	{
-		vscode.window.showInformationMessage('BOJ-EX를 설치해주셔서 감사합니다. 사용 전 설정을 해주세요.');
+		vscode.window.showInformationMessage('BOJ-EX를 설치해주셔서 감사합니다.');
 
+		showManual(context);
 		vscode.commands.executeCommand('workbench.action.openSettings', 'BOJ-extension');
 	}
 
@@ -84,6 +86,14 @@ export function activate(context: vscode.ExtensionContext)
 		vscode.commands.registerCommand('BOJ-EX.makeWorkflow', () =>
 		{
 			makeWorkflow();
+		}
+		));
+
+	// showManual 커맨드 등록
+	context.subscriptions.push(
+		vscode.commands.registerCommand('BOJ-EX.showManual', () =>
+		{
+			showManual(context);
 		}
 		));