+++ CHANGELOG.md
@@ -1,9 +1,28 @@
 # Change Log
 
-All notable changes to the "BOJ-EX" extension will be documented in this file.
+## [1.0.0] - 2023-08-02
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+### Added
+- Command: 문제 번호 입력 시 문제 검색
+- Command: 문제 번호 입력 시 헤더 자동 생성
 
-## [Unreleased]
+## [1.0.1] - 2023-08-02
 
-- Initial release
+### Added
+- Command: 문제 번호 입력 시 파일 생성, 문제 검색, 헤더 자동 생성
+
+## [1.0.2] - 2023-08-02
+
+### Added
+- 확장 아이콘 추가
+
+## [1.0.3] - 2023-08-03
+
+bug fix
+
+## [1.1.0] - 2023-08-08
+
+### Added
+- Command: Github Action을 위한 workflow.yml 생성 및 Github 푸시
+- Command: workflow.yml 생성
+- Command: 매뉴얼 열기